Use parking_lot Mutex where possible
parking_lot is a faster and more convenient alternative to std::sync::Mutex.
We can remove the annoying expect calls, as we don't care about it being poisoned.

It is used by many 3rd-party crates we use already, so it's not an additional dependency.
In addition, we can enable deadlock detection with this library
(done in a follow-up commit, just in case it adds too much overhead).

rust-lightning needs std::sync::Mutex, so leave these instances unchanged.