fix: Stabilise key dependencies
The main motivation is to stay on top of the latest changes to
`rust-lightning` and `rust-dlc`, whilst also definitively removing
some of the things that we temporarily added during our latest bug
hunt.

Some patches were also removed which might cause some controversy:

- https://github.com/get10101/rust-dlc/commit/ce78c25.
- https://github.com/get10101/rust-dlc/commit/cf3d1eb.
- https://github.com/get10101/rust-dlc/commit/304e85d.

In general reviewers are encouraged to take a look at the differences
between these branches:

- For `rust-dlc` we are now using
https://github.com/get10101/rust-dlc/commit/8fa3692 instead of
https://github.com/get10101/rust-dlc/commit/2322288.

- For `rust-lightning` we are now using
https://github.com/get10101/rust-lightning/commit/c65cb1c5 instead of
https://github.com/get10101/rust-lightning/commit/939ebb43.

The differences are not so obvious because we diverged quite a bit. We
still want to keep certain commits that haven't been upstreamed yet.

Other notable changes
---------------------

- Adapting to the new API to force-close the LN-DLC channel provided
by `rust-dlc`. As such, we've modified some code around that and
updated the `force_close_ln_dlc_channel` test accordingly.

- Updating `LnDlcNodeSettings` to include a couple more intervals so
that we can effectively disable certain tasks which can blow up when
we mine lots of blocks in certain tests, such as
`force_close_ln_dlc_channel`.

- Ensuring that we drop the settings `RwLock` guard when sleeping. We
were not doing so, so it was basically impossible to update the
`LnDlcNodeSettings` in the tests.

- Introducing `wait_until_dlc_channel_state` to replace a lot of eager
sleeps which led to slower and flakier tests.