fix: Stabilise key dependencies
The main motivation is to stay on top of the latest changes to
`rust-lightning` and `rust-dlc`, whilst also definitively removing
some of the things that we temporarily added during our latest bug
hunt.

Some patches were also removed which might cause some controversy:

- https://github.com/get10101/rust-dlc/commit/ce78c25.
- https://github.com/get10101/rust-dlc/commit/cf3d1eb.
- https://github.com/get10101/rust-dlc/commit/304e85d.

In general reviewers are encouraged to take a look at the differences
between these branches:

- For `rust-dlc` we are now using
https://github.com/get10101/rust-dlc/commit/8fa3692 instead of
https://github.com/get10101/rust-dlc/commit/2322288.

- For `rust-lightning` we are now using
https://github.com/get10101/rust-lightning/commit/c65cb1c5 instead of
https://github.com/get10101/rust-lightning/commit/939ebb43.

The differences are not so obvious because we diverged quite a bit. We
still want to keep certain commits that haven't been upstreamed yet.

Other notable changes
---------------------

- Adapting to the new API to force-close the LN-DLC channel provided
by `rust-dlc`. As such, we've modified some code around that and
updated the `force_close_ln_dlc_channel` test accordingly.

- Introducing `wait_until_dlc_channel_state` to replace a lot of eager
sleeps which led to slower and flakier tests.

- Enabling the configuration of the
`dlc_manager_periodic_check_interval` and
`sub_channel_manager_periodic_check_interval` viat the
`LnDlcNodeSettings`. Originally this was added to be able to
circumvent another issue, but the issue has now been fixed in
`rust-dlc`.

- Ensuring that we drop the `LnDlcNodeSettings` read lock as soon as
possible so that we don't hold onto it during the sleeps of certain
periodic tasks. This fixes a bug as in a previous version of this
patch we were not able to update the settings in the tests, because
there was always at least one read lock being held!