chore(ln-dlc-node): Improve ergonomics of load test
It cannot hang forever anymore and we are now able to see why posting
a trade is failing.

It is expected that trade attempts will fail a few times before
succeeding as we wait for the remote to update its internal state.
This is why we log on the failure on debug and why the message implies
that the failure is expected to be transient.