feat: Use address caches in LnDlcWallet
By caching these addresses we are preventing situations where getting
an address from the wallet takes a long time because the on-chain sync
is ongoing (and therefore has the mutex write lock).

The downside is that we can end up reusing addresses if these caches
are accessed more than once during a sync.