fix: Prefer unused addresses to new ones (temporarily)
AFAIK generating a new address means that the on-chain wallet must
look for that address on the blockchain. Thus, we must only generate a
new address if we are sure that we are about to use it.

As such, it is a bad idea to generate new addresses periodically. This
would quickly become untenable, as the on-chain sync would grow more
and more expensive and slow without good reason.

At the same time, it is best to use new addresses as it ensures that
we don't run into address reuse, which is a bad practice as it leads
to much worse privacy on-chain. Unfortunately, due to current
limitations of `bdk::Wallet` it is not possible to ask the wallet for
a new  (or even an unused) address during the on-chain sync. For this
reasons we decided to introduce an address cache.

Given all this, it's bad to keep a new-address cache so it has been
removed.

Furthermore, given that calls to
`get_address(AddressIndex::LastUnused)` already generate new addresses
if necessary and that we've already accepted that we are okay with
using (probably) unused addresses by introducing the address cache, we
have decided to _only_ allow calls to
get_address(AddressIndex::LastUnused). This simplifies the code and
ensures that we can always rely on the only cache we want to maintain
i.e. we never have to wait for the wallet to finish to get an address.