fix: Do not run on-chain sync in async context
The on-chain sync is blocking and it takes a long time. Such an
operation is not meant to be called from an async context directly. As
such, marking `ldk_node_wallet::Wallet::sync` as async is dangerous,
as any caller might end up blocking an async task unexpectedly.

We considered simply `spawn_blocking` the blocking call, but
`bdk::Wallet` is not `Send` as we already know.

As a side-effect, we no longer have to create a separate runtime to
run the on-chain sync. This runtime was kind of pointless, as it was
only ever scheduling one task at a time: the on-chain sync.

This change also makes it more obvious that the `ln-dlc-node`
`tokio::test`s are using `sync` in a dangerous way.