chore: Ensure coordinator syncs during `just fund`
If the coordinator misses a sync, then we cannot send funds to it, making it
harder to test the app (and resorting to workarounds, e.g. restarting it).

Ensure the coordinator is synced and has funds during funding.

The new endpoint (/api/admin/sync) should be used primarily for dev and testing.