chore: Add framework for end to end tests
End-to-end tests use the crate 'native' (the Rust-backend) as-is, without the
Flutter dependency.

Changes to `native` crate are minimal, mostly exposing types that were needed
and ensuring that we can link against it for tests (`rlib` target).

The tests live in a separate crate to ensure that we're only calling public APIs.

TestSubscriber ensures that we receive all the events destined for the GUI and
exposes an API to inspect current values for wallet info, position etc.

The end-to-end test needs to be run with `just e2e` command, to ensure
that the necessary components are up and the coordinator is funded.

The vision is that the coordinator and the maker instance will be shared across
all the tests and that each test will spawn its own app that can be run in parallel.

The sample test ensures at we can fund the app with a LND faucet.