fix(ln-dlc-node): Set ln::peer_handler logs to DEBUG in tests
These logs should be sufficient on DEBUG. On TRACE we log extremely
long lines every time we exchange CET signatures.

It's important to note that the more specific directive must come
_first_. Originally, we were setting `lightning=trace` first and then
attempting to set `peer_handler=debug`, but the latter was not taking
effect.

We also remove the `lightning::chain::channelmonitor=trace` directive
since it's superfluous.