fix(app): Remove print-logs button
It's a neat idea, but it doesn't work as implemented: we run into
extreme slowdowns if we have generated enough logs already, the log
window is often too small to see anything and it can be really slow to
scroll through the logs too.

For now we can make do with the share-logs button.