fix: Persist position and trade after dlc has been proposed
Fixes https://github.com/get10101/10101/issues/537

It is more likely that the proposing a dlc will fail. e.g. the oracle is not reachable, the peer lost connection. Hence, it's better to firstly try to propose the dlc and only when this is successful persist the position and trade to the database.