fix: Replace fee rate fallbacks in settings with constant
The fallbacks were only ever used if we could never get an answer from
the fee rate server. As such, it didn't make sense for them to be
configurable at runtime.

We move them to reduce complexity and because there is minimal benefit
in being able to configure values that will very rarely be used.