fix(app): Fix constant flickering on wallet screen
This was caused by calling `isUserSeedBackupConfirmed` (the future
passed to `FutureBuilder`) in `State<WalletScreen>.build`, something
that the docs explicitly say we must not do.

As recommended, we call it in `State<WalletScreen>.initState`. This
means that the `WalletScreen` is no longer informed about the updated
value after confirming that the seed has been backed up.

The simplest way to know is to return a value after
`pop`ping (returning from) the `SeedScreen`. This functionality is
only available in a later version of `go_router` so we update the
dependency to the latest version. With the returned value we can
update the dependency of `isUserSeedBackupConfirmed`.

Finally, we move the responsibility of keeping the button always
visible to the place where we actually enable the button. As such, we
don't need to comment out code and the TODO can be moved next to the
code that it references.