feat: Pay order-matching fee on the app
We keep it simple for now: the coordinator returns an invoice and we
trust the app to pay for it.

Invoice generation rationale
----------------------------
We choose to let the coordinator generate
an invoice rather than send an spontaneous payment from the app so
that we can more easily update the coordinator fees without having to
release a new version of the app.

Payment timing rationale
------------------------

The app pays the invoice as soon as the coordinator responds to the
trade request successfully. For now we let the coordinator claim the
payment immediately, but we could wait until the trade has actually
been executed to _claim_ the payment.

Claiming the payment later should be simple enough: let the
coordinator remember the `PaymentHash` of the order-matching fee
invoice when generating it and do not claim it automatically when
receiving the corresponding `PaymentClaimable` event (store the
preimage instead). Once the trade has been executed, look for the
invoice in storage and finally claim the funds.

Other thoughts
--------------

Eventually, the order-matching fee could be included in the payout
curve or, perhaps even better, as a separate static output on the
buffer transaction (this is currently not supported).