Defocus textfield upon clicking anywhere on send screen
The `GestureDetector` records taps within the `SafeArea` of the screen.
Setting the `FocusScope` to an unused `FocusNode` removes the focus from the `TextField`.
`behavior: HitTestBehavior.opaque,` allows clicking in "empty" parts of the screen; otherwise the focus would only be take away from the `TextField` when e.g. clicking on other text.