chore: Remove custom database Docker image
Changes in `database/init.sql` are not reflected unless the database container
is removed, which does not happen by default.

This caused recent addition of `oracle` table to not work at anyone who had
previously used this container.

It's enough to mount the script at runtime to a vanilla container, so remove the
custom Dockerfile altogether.

As a bonus, we should get slightly better build times on our tests, as we're not building
a container anymore, but using a cached one directly.