Coordinator creates channels based on user defined amounts
Instead of using fixed `200k` channels the coordinator create channels based on the liquidity that the user provides.
The coordinator multiplies the provided liquidity with `LIQUIDITY_MULTIPLIER`.
We define a maximum channel value of `JUST_IN_TIME_CHANNEL_OUTBOUND_LIQUIDITY_SAT_MAX` (remains at `200k`).
Given that `LIQUIDITY_MULTIPLIER` is currently configured to `2` this means that a user requests to open a channel with liquidity higher than `100k` will be rejected by the coordinator.