feat(app): Pay order-matching fee _after_ trade is executed
This patch includes an explicit sleep that should be removed once we
understand why we can fail to emit the `PaymentClaimable` event in the
coordinator and we implement a fix.

Another caveat is that the order-matching fee invoice is stored in
memory, so we could theoretically end up closing the app before the
payment goes through.