Calculate and show channel opening fees including modal info
If we don't have a `ChannelInfo` yet we know this is an initial channel open and estimate the channel opening fee.
The fee information is based on a transaction weight estimate of `220` for the funding transaction estimate with two inputs. For most users this will likely be cheaper because only one UTXO will be used but it's better to safe funds than to have to pay more.

If we have a fee on record when creating an invoice we add this fee to the minimum receive value to ensure the user will be able to pay the fee and trade.
We pass a potential fee on to the share invoice screen where the fee is displayed including an info button that pops up an explanatory modal.