feat: Shadow channel data
This allows us to keep channel data even if the channel is in pending or closed. Additionally, we store meta-data e.g. describing who initiated the force-closure of a channel.

As we will later on have to store additional transaction like commitment tx, split tx, etc. shadowing the ldk channel is the only way to not lose this data.