feat: Configure proportional routing fee consistently
- Align the value of the proportional routing fee used when (1) giving
data to the app to create an interceptable invoice to open a JIT
channel and when (2) opening a channel as a result of intercepting a
payment.

- Allow the coordinator to configure the proportional routing fee via
the config file and through the update-settings endpoint at runtime.
In theory, this means that the values could vary between the two
scenarios described above, but I see that as a feature.

- Ensure that the coordinator always uses the up-to-date `UserConfig`
when opening new JIT channels and regular public channels.

These changes do NOT ensure that:

- Inbound channels will respect a `Node`'s `UserConfig` updated at
runtime. This is because LDK does not currently support updating the
`ChannelManager::default_configuration` other than by restarting the
`ChannelManager` with a new configuration[1].

- Existing channels have their settings updated every time the
coordinator's config changes. This should be possible by calling
`Channel::update_config` on all open channel.

Additionally:

- Use to_ as prefix where suitable. The `as_` prefix is reserved for
borrowed-to-borrowed conversions[2].

- Give control of last route hint hop to coordinator.

Before we were just forwarding part of the last route hint hop
information to be embedded in the interceptable invoice.

With this change we are now able to control everything (including the
`base_msat` fee which was updated recently and would have been out of
sync).

[1]: https://github.com/lightningdevkit/rust-lightning/issues/2420.
[2]:
https://rust-lang.github.io/api-guidelines/naming.html#ad-hoc-conversions-follow-as_-to_-into_-conventions-c-conv.