fix(ln-dlc-node): Fail intercepted HTLC backwards always if needed
All failures to handle the `HTLCIntercepted` event should result in
calling `ChannelManager::fail_intercepted_htlc` to fail the HTLC
backwards. Failing to do this means that the HTLC will be pending for
a very long time, negatively impacting all channels in the path.

By defining a fallible function `handle_intercepted_htlc_internal`
which simply bubbles up errors and a function
`handle_intercepted_htlc` which fails the HTLC backwards if
`handle_intercepted_htlc_internal` returns an error we ensure that we
always fail the HTLC backwards if necessary.