fix(ln-dlc-node): Fail intercepted HTLC on ChannelClosed event
We do not try to be smart because it's not necessary: if there is a
corresponding entry in the `PendingInterceptedHtlcs` map we assume
that the HTLC is still pending and fail it accordingly.

I believe we currently do not clear the map after forwarding the
intercepted HTLC, but this is irrelevant as failing an intercepted
HTLC that was already forwarded (or that was already failed) is a
no-op.