feat: Manage spendable outputs generated by LDK
For all `lnd_dlc_nod::Node`s we spin up a task which ensures that LDK
spendable outputs that are "owned" by LDK are moved to the `Node`s
on-chain wallet.

This functionality could be split between coordinator and app (the two
types of `Node` currently conceived), but it does not seem worth it at
this stage.