feat: Track potential JIT channels in coordinator DB
This should allow us to see how often app users want to open JIT
channels but fail to do so.

It also makes it easier to track the path from registering interest in
opening a JIT channel to actually opening the JIT channel. That should
help with debugging.

This patch also highlights some of the pains of using the same shadow
`Channel` model for both coordinator and app. Here we introduce
changes that are definitely only for the coordinator, but we end up
needing to update the app's DB `Channel` model.

Co-authored-by: Daniel Karzel <daniel@comit.network>