feat: Update to latest rust-dlc
This includes the long-awaited feature of being able to force-close an
LN-DLC channel in every intermediate state.

The notable changes are:

- Add new `Revoke` message to the subchannel-opening protocol. This
has forced us to adapt the tests. More importantly we now update the
position and order of the _app_ when receiving this message, rather
than when sending the `Finalize` message. This is because the `Revoke`
message is what actually marks the end of the subchannel-opening
protocol now[1].

- Require consumers to notify the `SubChannelManager` about the
closure of a channel according to LDK. This is necessary so that the
subchannel manager can react to the counterparty force-closing the
channel using an untracked transaction.

Finally, by now depending on latest `rust-dlc#feature/ln-dlc-channels`
we have replaced
https://github.com/lightningdevkit/rust-lightning/commit/a663e1f with
https://github.com/lightningdevkit/rust-lightning/commit/74690be.

[1]: To be fair, without the `Revoke` message both versions of the
LN-DLC channel are valid: one with the subchannel channel and one
without. This is because the coordinator can still commit to the old
version without the subchannel. Since the app doesn't currently
attempt to convey this information to the user, it seems simpler to
default to only updating this data after the protocol is complete.