fix(app): Wait longer for capacity to pay for channel-opening fee
Waiting for 5 seconds proved insufficient in some cases. In theory we
could wait for as long as we want to as this work will happen in the
background.

It might still help to redesign this flow so that we just keep
retrying if we owe the coordinator for opening the JIT channel, rather
than only trying once after claiming the interceptable payment.