fix: install rust nightly and dprint
Same as in commit c38dd5721dfb1e7e06ab5640f2f817c9d284883b we need a small workaround to get dprint working in CI.

Using the github action with check is a bit of an overhead but ensures that we are installing the same dprint version as elsewhere.
We need to install said dprint version before we do any changes as the check might fail otherwise.