chore(justfile): Don't invalidate the cache between iOS and native builds
It was a major slowdown to wait for the project to recompile as cache was being invalidated.

By specifying a different target directory we ensure build artefacts will be
preserved when switching between `just ios` and `just native`.

Note: Only the debug build has been adjusted, as this should not matter for the releases.