feat: Display channel status in app
There are many reasons why the app might end up with a channel
(vanilla or split) closed on-chain, particularly at this stage of the
project.

As such we want to give some visibility to the app user, so that they
know what to expect if they run into this situation. The main
objective is to prevent loss of funds, something that could happen if
the user deletes the app data before certain transactions have been
confirmed.

Interestingly, for the DLC channel branch of the split channel, the
user could choose to trust 10101, as the coordinator will eventually
publish the correct CET, partially paying to the user's on-chain
wallet.

The LN branch is more dangerous, as the user must wait for the funds
to be claimed from the commitment transaction by their own LN node.
Since we currently do not back up this data, deleting the app data at
the wrong time could lead to loss of funds.

The main limitation of this patch, captured in a TODO, is that we
currently do not model the LN _closing_ channel state in our shadow
channel representation. This means that we will end up telling the
user that the LN channel is closed before it is fully closed, which is
dangerous.

The user can also use the channel status information to know if they
are able to open a new channel _with the same app_. If the channel
status is `NotOpen` they can create a new JIT channel. This
information is useful because they will know that they can continue
using 10101 whilst their previous channel might still be being closed
in the background.

Also:

- Register new incoming JIT channels with a `UserChannelId`. This is
important for this patch because otherwise we cannot identify the
channel, which prevented us from marking it as closed in its shadow
representation.