chore: print pods runner script
Locally I have the problem that the script is being manipulated when running fastlane and a crucial line is being changed which makes the command fail.
Locally I have to watch the build and at the right moment have to edit the file on line 44 to `source="$(readlink -f "${source}")"`, i.e. add the `-f`.

This patch just prints the file to check if CI also suffers of the same bug.