fix(test-e2e): Wait for app outbound payment after opening channel
Fixes https://github.com/get10101/10101/issues/1087.

We fix two issues here:

- We no longer assume that the order-matching fee is constant. It
actually depends on the state of the blockchain, which can change as
we run the e2e tests (even if run sequentially!).
- We wait for the app to confirm the payment before continuing with
test assertions. Otherwise these tests can be flaky.

The fact that the test needs to know about certain implementation
details (e.g. that the order-matching fee is paid via a payment after
the JIT channel is open) is not great, but it's pretty hard to know
what fee will be charged from within the test before it is actually
paid.

This test will probably have to change when we address
https://github.com/get10101/10101/issues/843.