feat: Add Lightning-node capabilities to the maker
Fixes https://github.com/get10101/10101/issues/926.

The main goal of this patch is to give the maker the ability to open a
channel with the coordinator to make direct payments. These will
firstly be used to pay order-matching fees. Eventually they will also
be used to set up DLC channels corresponding to trades.