feat: Check if DLC message queue is empty before getting
At the moment it's not a big deal because, knowing the implementation
details of `DlcMessageHandler`, it doesn't make a big difference.
Doing this we currently do allocate unnecessarily if we call
`get_and_clear_received_messages` and there are no messages.

To be able to do this we have to update `rust-dlc` to this patch:
https://github.com/p2pderivatives/rust-dlc/commit/22b7052. We also
update `rust-lightning` to match the version in `rust-dlc`.