fix: Add buffer delay after outbound capacity is ready
It could happen that the channel already shows outbound capacity due to the received payment, but the channel updates are not completely finished yet. We suspect that this could lead to issues when trying to pay the JIT channel opening fees at the same time. Resulting in force closures or worst case and unresponsive app (deadlock?).

By waiting additional 500 milliseconds after the outbound capacity is set, we hopefully prevent paying an invoice while the channel is still updated.