feat: Show loading screen when app starts with and expired position
This commit introduces the order reason and the force match. Basically the idea is that we have certain orders that will have to get created by the orderbook (on behalf of the user like expiry, liquidation, take profit and stop loss). Those orders are matched automatically without the users intervention. Once matched we inform the user about the match for execution. If the user is online the trade will be automatically executed if not, the execution is delayed until the user comes online. Note, this match can still expire.