chore: Separate components into separate modules
Before everything was handled by the trading module. Now the following responsibilities have been split up.

- Rollover: Moved to the coordinator component and responsible for proposing a rollover.
- Async Match: Moved into a dedicated component to check if an async match needs to be executed by the app.
- Notification: Responsible for the users and sending messages to them.

Note 1, the websocket communication is still in one component and not separated between coordinator and orderbook. It would have been correct to split that up as well, but the effort and additional complexity was IMHO not worth it.

Note 2, we have multiple commons crates, which are very much related to each ohter. I think we should combine all of those into a single one, to simplify things.