feat(app): Delay first periodic on-chain sync by a minute
To ensure that the wallet refresh can happen quickly before the first
on-chain sync. This allows the wallet to appear ready before finishing
the first on-chain sync.

To achieve this we've had to remove the automatic on-chain sync from
the "default" version of `ln_dlc_node::Node`. Now consumers have to
set up the periodic task themselves, although the library provides
convenience methods.