feat(app): Do not force on-chain sync on startup
We already have:

- A periodic on-chain sync starting one minute after the app has
started.
- A periodic wallet refresh which updates the wallet history and
balances every 5 seconds (without syncing on-chain every time).

To have an app which appears ready as soon as possible, what we
actually want is to refresh the wallet balances and history as soon as
possible and delay the on-chain sync until a bit later. This is
already achieved by the periodic tasks described above, but we need to
remove a manual call to `refreshWalletInfo`, because that one actually
triggers the on-chain sync.

We are obviously introducing a trade-off, as the user will no longer
see the latest state of their wallet in terms of the blockchain as
soon as the app appears ready.