chore: Increase websocket reconnect timeout
If we are reconnecting too eagerly we run into an issue during rollover windows.

If the user has the app open while "moving" into the rollover window and the coordinator restarts. The coordinator will identify that the users position needs to be rolled over upon a reconnect on the websocket, but the coordinator might not yet be ready to do that yet.

I guess this issue is not happening on the app side, as the app is connecting to the coordinator before it is subscribing to the websocket.

By increasing the websocket reconnect timeout to 5 seconds the `keep_connected` tasks should have enough time (every second) to reconnect to the coordinator.