fix: Do not delete orders
Instead of deleting orders, I propose to keep them in state `Failed`. That might blow up our database, but we could at any time delete `Failed` orders if we want to.

To be consistent with the existing implementation I am also sending a `DeleteOrder` event to update the prices for the apps.