feat: In-app waitlist
Check access to the waitlist during in-app registration and prevent from using
the app if they're not allowed.

- If the user is not on the waitlist, add them and display a message that they
  do not have access yet, pointing to our Telegram as well.
- if the user is on the waitlist, check whether they're allowed to access.

Only after successful admission store the email and allow

Add a coordinator setting to allow auto-adding new users (to allow throttling
users during open beta, allowing us to stay within our liquidity constraints).

Add an API to the coordinator to query the waitlist and to add a user to the
waitlist (with options to enable/disable access).

The waitlist is stored as a `waitlist` table in coordinator database, initially
populated from the `users` table to ensure all our current users have immediate access.

Note: I didn't try to go too fancy initially. In the futurue, we could e.g.
     unsubscribe from the orderbook stream if we just want to disable the
     trading functionality.

block the access to the app for unregistered users.