feat: flavor dependent firebase config (iOS)
In this patch we copy firebase configs (GoogleService-Info.plist) dependent on what flutter flavor is defined. This is needed because we have two different bundle IDs for production vs testnet and a singel firebase project can't handle this.
This was implemented based on the blog post here: https://ahmedyusuf.medium.com/setup-flavors-in-ios-flutter-with-different-firebase-config-43c4c4823e6b