fix(app): Ensure match between opening/closing trades in history
By rounding the order quantities (contracts) to 2 decimal places it is
_much less likely_ that opening and closing trades do not match.

A mismatch can break the trades in the wallet and trade history
view, because a closing trade can  end up being ignored as an
unsupported (extending, reducing, direction-changing) trade. This
error can compound, as the error is carried forward an all subsequent
trades will almost certainly be missing too.

Once we add support for all other forms of trading, this error can
mostly go away as all cases will be handled.