fix(fund): Don't check for peer alias by default
We can already check whether the channel we have just opened is active; this
check is likely not needed anymore. For the cases where it is desired, keep it
with an opt-in flag.

Broadcasting peer announcements, especially on the local channel, has proven to
be unreliable. This in turn caused errors when running `just fund` locally.