feat: Recover from stuck rollover state
- Rollback channel state to the last stable state if in intermediate state at reconnect
- Adds the rollover contract to the rollover proposal

Upon starting the app it checks if the channel is in an intermediate state. If so it rolls back to the last "stable" state. The coordinator does the same upon connecting with the app and will the propose another rollover.

Note, in case the app broke before processing the `RenewRevoke` message, the coordinator will not be able to identify that the rollover failed for the user and remains the renewed contract. Meaning the rollover will not be retried. That puts the coordinator at an advantage and is just another reason why this bug should be fixed properly in rust-dlc eventually. However, the channel and subchannel is usable for both parties and a collaborative close is possible. If the renewed contract on the cooridinator side were to expire again, the rollover would be rejected by the app, as it is in the context of a contract unkown to the app. So the user can either close the position or wait for it to expire.

Note, this is merely a quick fix to unblock users that are in a stuck state after rollover. A proper fix should be implemented in rust-dlc, rendering this fix obsolete.