fix: Notify users on websocket if online for rollover
If the app is in background it might still be connected to our websocket. In that case we don't have to notify the user via push notification, but can directly send the websocket message.

Note, it might happen that the user has the app in the background and does not open the app again at the time of notification. In that case the app would disconnect eventually, the app wouldn't have rolled over (unless we are very lucky and the app is not in the background for long) and the user would have missed that push notification. However, we are sending atm 3 push notifications in total during the rollover weekend, so that should be fine.