chore: Upgrade to rust-lightning 0.0.116 and bump rust-dlc
Some notes:

- We've had to increase the payment amounts in the `multi_hop_payment`
and `just_in_time_channel_with_multiple_payments` tests, as otherwise
we would run into `route not found` errors. It's probably worth
investigating why this is the case.

- We have decided to eagerly send `rust-dlc` messages because: (1)
after the upgrade they were not being automatically sent as fast as
they used to and (2) we can speed up the protocols this way anyway.

Co-authored-by: Philipp Hoenisch <philipp@hoenisch.at>